implementação de multer

+++ src/modules/User/useCases/delete/DeleteUser.service.ts
@@ -1,6 +1,7 @@
-import AppError from "http/error/AppError"
-import { IUsersRepository } from "modules/User/repositories/IUserRepository"
-import { inject, injectable } from "tsyringe"
+import { IUserResponseDTO } from 'modules/User/dtos/IUser.dto'
+import { IUsersRepository } from 'modules/User/repositories/IUserRepository'
+import { inject, injectable } from 'tsyringe'
+import AppError from '../../../../http/error/AppError'
 
 @injectable()
 export class DeleteUserService {
@@ -9,10 +10,19 @@ export class DeleteUserService {
         private userRepository: IUsersRepository
     ) {}
 
-    async execute(id: number): Promise<Boolean> {
+    async execute(id: number, userLogged: IUserResponseDTO): Promise<Boolean> {
+        if (userLogged.id !== id && userLogged.groupUserId !== 1) {
+            throw new AppError('Usuário não tem permissão para essa ação', 401)
+        }
+
         const userExists = await this.userRepository.findById(id)
         if (!userExists) throw new AppError('Usuário não encontrado.', 404)
-        await this.userRepository.delete(id)
-        return
+
+        try {
+            await this.userRepository.delete(id)
+            return true
+        } catch (error) {
+            return false
+        }
     }
 }
